+++ resources/views/examples/countdown/default.blade.php
@@ -1,0 +1,1 @@
+<x-ui.countdown :to="now()->addDays(12)->addHours(7)->addMinutes(32)->addSeconds(8)" />